pyproject.toml
[project]
name = "amazon-products"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "load-dotenv>=0.1.0",
    "pandas>=2.3.1",
    "python-amazon-paapi>=5.0.1",
    "xlrd>=2.0.2",
]
